New 0.11 Return
I haven't been developing this for a while for I though Typescript could be a better solution than Lua in Warcraft III scripting. However I found that Typescript is not only complex, but also unstable because of the translation. So I've decided to make a return to Lua.
All you see at this stage are what I wish to present in the release version, so they may contain scripts from other authors on HIVE. They shall be replaced with mine before being released.